Refacto screencast widget
Before:
- screencast widget was using a single button for both screenshot and screenrecord
- Time of screenrecord was displayed near toolbar in a tooltip

After:
- screencast widget deleted
- options to activate screenshot and screenrecord are still 'capture' (option activated or desactivated both plugins)
- screend record and scren shot are now two separate button in the toolbar
- time of screenrecord is now displayed in the toolbar (below the record button)
- when screenrecord is active, the background of the player is changed

Screenrecord and screenshot buttons can be included in the toolbar or in the floating toolbar
Adding tests for screenshot and screenrecord buttons
