+++ README.md
@@ -451,14 +451,14 @@ A device renderer instance can be configured using the `options` argument (objec
 -   **Details:**
     Enable or disable the clipboard widget. This widget allows you to share clipboard content between the local machine and the Android virtual device, enabling seamless copy-paste functionality in both directions.
 
-#### capture <img align="right" src="./doc/assets/ic_screencast.svg">
+#### capture <img align="right" src="./doc/assets/ic_screenshot.svg"><img align="right" src="./doc/assets/ic_screenrecord.svg">
 
 -   **Type:** `Boolean`
 -   **Default:** `true`
 -   **Toolbar name:** `Screencast`
 -   **Compatibility:** `PaaS`, `SaaS`
 -   **Details:**
-    Enable or disable the capture widget. This widget can be used to capture the screen of the Android virtual device (screenshot or screencast).
+    Enable or disable the capture widget. This widget adding two button to the toolbar for screenshot or screenrecord.
 
 #### diskIO <img align="right" src="./doc/assets/ic_diskIO.svg">
