--- src/scss/components/_screencast.scss
@@ -1,103 +1,0 @@
-/**
- * Screencast plugin styles
- */
-.device-renderer-instance {
-    .gm-screencast-plugin {
-        // TODO delete this line in the PR which will refacto this plugin, keep for css compatibility
-        label{
-            color: var(--gm-text-color);
-            font-weight: normal;
-        }
-        width: 400px;
-        left: calc(50% - 200px);
-
-        /* Extra Small Devices, Tablet or Phones */
-        @media only screen and (max-width: $breakpoint-mobile) {
-            max-width: 400px;
-        }
-
-        label {
-            height: 32px;
-            line-height: 35px;
-            margin-bottom: 0;
-            margin-left: 20px;
-            cursor: pointer;
-        }
-
-        .disabled {
-            opacity: 0.4;
-            pointer-events: none;
-        }
-
-        .gm-inputs {
-            display: flex;
-        }
-
-        .gm-screenshot,
-        .gm-screencast {
-            display: inline-block;
-            width: auto;
-            margin-right: 50px;
-            display: flex;
-            flex-direction: row-reverse;
-
-            .gm-action {
-                width: 32px;
-                min-height: 32px;
-                cursor: pointer;
-                display: block;
-                -webkit-mask-size: contain;
-                mask-size: contain;
-                -webkit-mask-repeat: no-repeat;
-                mask-repeat: no-repeat;
-                -webkit-mask-position: center;
-                mask-position: center;
-                background-color: var(--gm-on-secondary-color);
-            }
-        }
-
-        .gm-screenshot .gm-action {
-            mask-image: url('../assets/images/ic_screenshot.svg');
-            --webkit-mask-image: url('../assets/images/ic_screenshot.svg');
-        }
-
-        .gm-screencast .gm-action {
-            mask-image: url('../assets/images/ic_screencast.svg');
-            --webkit-mask-image: url('../assets/images/ic_screencast.svg');
-
-            .disabled {
-                color: var(--gm-btn-bg-color-disabled);
-            }
-        }
-    }
-
-    .gm-screencast-button:hover {
-        + .gm-screencast-timer {
-            display: block;
-        }
-    }
-    .gm-screencast-timer {
-        position: fixed;
-        right: 57px;
-        color: #c6225a;
-        font-size: 20px;
-        font-family: Helvetica, sans-serif;
-        pointer-events: none;
-        padding: 11px 8px 8px 8px;
-        border-radius: 6px;
-        width: 80px;
-        background-color: #303339;
-        margin-top: -37px;
-        display: none;
-
-        &::after {
-            content: ' ';
-            position: absolute;
-            right: -10px;
-            top: 16px;
-            border-width: 5px;
-            border-style: solid;
-            border-color: transparent transparent transparent #303339;
-        }
-    }
-}
