+++ src/scss/main.scss
@@ -13,7 +13,7 @@
 @import 'components/_upload';
 @import 'components/_battery';
 @import 'components/_gps';
-@import 'components/_screencast';
+@import 'components/_screenrecord';
 @import 'components/_imei';
 @import 'components/_network';
 @import 'components/_iothrottling';
